Prime Number
This is a simpler approach to decide whether a given number is prime or not by considering the nearest square possible to the given number. The range is initialized by the nearest square of number and basically sets itself as the limit and check if number is divisible by prime number in the given range as such.